// src/pages/group-operations/Report/Report.enum.ts
export enum EReportType {
   CurrentIncomesCosts = 'current-incomes-costs',
   ConsultProjects = 'consult-projects',
   CostsProjectedMonthly = 'consts-projected-monthly'
}

export enum EAreasReportIncomesCosts {
   Consultancy = 'consultoria',
   Catalyst = 'catalyst'
}

export enum EReportConsultProjects {
   Consolidated = 1,
   Detailed = 2
}

export enum EAreasReportCostsProjectedMonthly {
   Consultancy = 'consultoria',
   Catalyst = 'catalyst'
}